scorer/requirements.txt
python-arango==4.2.1
py_expression_eval==0.3.10
